Clean things up a bit
