+++ axum-liveview/src/pubsub/in_process.rs
@@ -19,19 +19,21 @@ impl Default for InProcess {
 }
 
 #[async_trait]
-impl PubSub for InProcess {
-    async fn send_raw(&self, topic: &str, msg: Bytes) -> anyhow::Result<()> {
+impl PubSubBackend for InProcess {
+    async fn broadcast_raw(&self, topic: &str, msg: Bytes) -> anyhow::Result<()> {
         let _ = self.tx.send((topic.to_owned(), msg));
         Ok(())
     }
 
-    async fn subscribe_raw(&self, topic: &str) -> BoxStream<'static, Bytes> {
+    async fn subscribe_raw(&self, topic: &str) -> anyhow::Result<BoxStream<'static, Bytes>> {
         let topic = topic.to_owned();
 
-        BroadcastStream::new(self.tx.subscribe())
+        let stream = BroadcastStream::new(self.tx.subscribe())
             .filter_map(|result| ready(result.ok()))
             .filter(move |(msg_topic, _)| ready(**msg_topic == topic))
             .map(|(_, msg)| msg)
-            .boxed()
+            .boxed();
+
+        Ok(stream)
     }
 }
