+++ axum-liveview/src/ws.rs
@@ -1,6 +1,5 @@
 use crate::{
     html::{self, Diff},
-    liveview::{EventContext, WithEventContext},
     pubsub::PubSub,
     topics, LiveViewManager,
 };
@@ -14,7 +13,7 @@ use axum::{
 use futures_util::{stream::BoxStream, StreamExt};
 use serde::{Deserialize, Serialize};
 use serde_json::{from_value, json, Value};
-use std::{collections::HashMap, fmt, time::Duration};
+use std::{collections::HashMap, time::Duration};
 use tokio::time::{timeout, Instant};
 use tokio_stream::StreamMap;
 use uuid::Uuid;
@@ -33,7 +32,6 @@ async fn ws(upgrade: WebSocketUpgrade, live: LiveViewManager) -> impl IntoRespon
 #[derive(Default)]
 struct SocketState {
     diff_streams: StreamMap<Uuid, BoxStream<'static, Diff>>,
-    js_command_streams: StreamMap<Uuid, BoxStream<'static, JsCommand>>,
 }
 
 async fn handle_socket<P>(mut socket: WebSocket, pubsub: P)
@@ -47,8 +45,7 @@ where
     const HEARTBEAT_FREQUENCY: Duration = Duration::from_secs(5);
     const HEARTBEAT_MAX_FAILED_ATTEMPTS: usize = 5;
 
-    // let mut heartbeat_interval = tokio::time::interval(HEARTBEAT_FREQUENCY);
-    let mut heartbeat_interval = tokio::time::interval(A_VERY_LONG_TIME);
+    let mut heartbeat_interval = tokio::time::interval(HEARTBEAT_FREQUENCY);
     let mut failed_heartbeats = 0;
     let mut heartbeat_sent_at = Instant::now();
 
@@ -115,19 +112,6 @@ where
             Some((liveview_id, diff)) = state.diff_streams.next() => {
                 let _ = send_message_to_socket(&mut socket, liveview_id, RENDERED_TOPIC, diff).await;
             }
-
-            Some((liveview_id, js_command)) = state.js_command_streams.next() => {
-                let msg = match js_command {
-                    JsCommand::NavigateTo { uri } => json!({
-                        "type": "navigate_to",
-                        "data": {
-                            "uri": uri,
-                        }
-                    }),
-                };
-
-                let _ = send_message_to_socket(&mut socket, liveview_id, JS_COMMAND_TOPIC, msg).await;
-            }
         }
     }
 
@@ -157,7 +141,6 @@ async fn send_heartbeat(socket: &mut WebSocket) -> Result<(), axum::Error> {
 const INITIAL_RENDER_TOPIC: &str = "i";
 const LIVEVIEW_GONE_TOPIC: &str = "liveview-gone";
 const RENDERED_TOPIC: &str = "r";
-const JS_COMMAND_TOPIC: &str = "j";
 
 async fn send_message_to_socket<T>(
     socket: &mut WebSocket,
@@ -193,13 +176,8 @@ where
         ($expr:expr, $pattern:path $(,)?) => {
             match $expr {
                 $pattern(inner) => inner,
-                other => {
-                    tracing::error!(
-                        file = file!(),
-                        line = line!(),
-                        ?other,
-                        "`tri!` didn't match"
-                    );
+                _other => {
+                    tracing::error!(file = file!(), line = line!(), "`tri!` didn't match");
                     return None;
                 }
             }
@@ -226,8 +204,10 @@ where
 
     match msg {
         EventFromBrowser::Mount => {
-            let mut initial_render_stream =
-                pubsub.subscribe(&topics::initial_render(liveview_id)).await;
+            let mut initial_render_stream = tri!(
+                pubsub.subscribe(&topics::initial_render(liveview_id)).await,
+                Ok,
+            );
 
             tri!(
                 pubsub.broadcast(&topics::mounted(liveview_id), ()).await,
@@ -246,28 +226,18 @@ where
                     }
                 };
 
-            let diff_stream = pubsub
-                .subscribe(&topics::rendered(liveview_id))
-                .await
+            let diff_stream = tri!(pubsub.subscribe(&topics::rendered(liveview_id)).await, Ok)
                 .map(|Json(diff)| diff);
             state
                 .diff_streams
                 .insert(liveview_id, Box::pin(diff_stream));
-
-            let js_command_stream = pubsub
-                .subscribe(&topics::js_command(liveview_id))
-                .await
-                .map(|Json(js_command)| js_command);
-            state
-                .js_command_streams
-                .insert(liveview_id, Box::pin(js_command_stream));
 
             return Some(HandledMessagedResult::Mounted(
                 liveview_id,
                 initial_render_html,
             ));
         }
-        EventFromBrowser::Click(Click { msg }) => {
+        EventFromBrowser::Click(ClickEvent { msg }) => {
             let topic = topics::update(liveview_id);
             let msg = WithEventContext {
                 msg,
@@ -275,7 +245,7 @@ where
             };
             tri!(pubsub.broadcast(&topic, Json(msg)).await, Ok);
         }
-        EventFromBrowser::FormEvent(FormEventMessage { msg, value }) => {
+        EventFromBrowser::FormEvent(FormEvent { msg, value }) => {
             let topic = topics::update(liveview_id);
             let msg = WithEventContext {
                 msg,
@@ -283,7 +253,7 @@ where
             };
             tri!(pubsub.broadcast(&topic, Json(msg)).await, Ok);
         }
-        EventFromBrowser::KeyEvent(KeyEventMessage {
+        EventFromBrowser::KeyEvent(KeyEvent {
             msg,
             key,
             code,
@@ -369,41 +339,27 @@ impl TryFrom<RawMessage> for EventFromBrowser {
 #[derive(Debug)]
 enum EventFromBrowser {
     Mount,
-    Click(Click),
-    FormEvent(FormEventMessage),
-    KeyEvent(KeyEventMessage),
+    Click(ClickEvent),
+    FormEvent(FormEvent),
+    KeyEvent(KeyEvent),
 }
 
 #[derive(Debug, Deserialize)]
-struct Click {
+struct ClickEvent {
     #[serde(rename = "m")]
     msg: Value,
 }
 
 #[derive(Debug, Deserialize)]
-struct FormEventMessage {
+struct FormEvent {
     #[serde(rename = "m")]
     msg: Value,
     #[serde(rename = "v")]
     value: FormEventValue,
 }
 
-#[derive(Serialize, Deserialize, Debug)]
-#[serde(untagged)]
-pub(crate) enum FormEventValue {
-    String(String),
-    Strings(Vec<String>),
-    Bool(bool),
-    Map(HashMap<String, FormEventValue>),
-}
-
 #[derive(Debug, Deserialize, Serialize)]
-pub(crate) enum JsCommand {
-    NavigateTo { uri: String },
-}
-
-#[derive(Debug, Deserialize, Serialize)]
-struct KeyEventMessage {
+struct KeyEvent {
     #[serde(rename = "m")]
     msg: Value,
     #[serde(rename = "k")]
@@ -418,6 +374,53 @@ struct KeyEventMessage {
     shift: bool,
     #[serde(rename = "me")]
     meta: bool,
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+pub(crate) struct WithEventContext<T> {
+    pub(crate) msg: T,
+    pub(crate) ctx: EventContext,
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+pub struct EventContext {
+    inner: EventContextInner,
+}
+
+impl EventContext {
+    pub(crate) fn click() -> Self {
+        Self {
+            inner: EventContextInner::Click,
+        }
+    }
+
+    pub(crate) fn form(value: FormEventValue) -> Self {
+        Self {
+            inner: EventContextInner::Form(value),
+        }
+    }
+
+    pub(crate) fn key(value: KeyEventValue) -> Self {
+        Self {
+            inner: EventContextInner::Key(value),
+        }
+    }
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+enum EventContextInner {
+    Click,
+    Form(FormEventValue),
+    Key(KeyEventValue),
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+#[serde(untagged)]
+pub(crate) enum FormEventValue {
+    String(String),
+    Strings(Vec<String>),
+    Bool(bool),
+    Map(HashMap<String, FormEventValue>),
 }
 
 #[derive(Debug, Deserialize, Serialize)]
@@ -436,51 +439,7 @@ pub(crate) struct KeyEventValue {
     pub(crate) meta: bool,
 }
 
-macro_rules! axm {
-    (
-        $(#[$meta:meta])*
-        pub(crate) enum $name:ident {
-            $(
-                #[attr = $attr:literal]
-                $(#[$variant_meta:meta])*
-                $variant:ident,
-            )*
-        }
-    ) => {
-        $(#[$meta])*
-        pub(crate) enum $name {
-            $(
-                $(#[$variant_meta])*
-                $variant,
-            )*
-        }
-
-        impl $name {
-            #[allow(warnings)]
-            pub(crate) fn from_str(s: &str) -> anyhow::Result<Self> {
-                match s {
-                    $(
-                        concat!("axm-", $attr) => Ok(Self::$variant),
-                    )*
-                    other => anyhow::bail!("unknown message topic: {:?}", other),
-                }
-            }
-        }
-
-        impl fmt::Display for $name {
-            fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-                match self {
-                    $(
-                        Self::$variant => write!(f, concat!("axm-", $attr)),
-                    )*
-                }
-            }
-        }
-    };
-}
-
 axm! {
-    #[non_exhaustive]
     #[derive(Debug)]
     pub(crate) enum Axm {
         #[attr = "blur"]
