+++ examples/counter/src/main.rs
@@ -1,5 +1,5 @@
-use axum::{async_trait, response::IntoResponse, routing::get, Json, Router};
-use axum_liveview::{html, liveview::EventContext, Html, LiveView, LiveViewManager, Setup};
+use axum::{async_trait, response::IntoResponse, routing::get, Router};
+use axum_liveview::{html, EventContext, Html, LiveView, LiveViewManager, Subscriptions};
 use serde::{Deserialize, Serialize};
 use std::net::SocketAddr;
 
@@ -52,7 +52,7 @@ struct Counter {
 impl LiveView for Counter {
     type Message = Msg;
 
-    fn setup(&self, setup: &mut Setup<Self>) {}
+    fn init(&self, subscriptions: &mut Subscriptions<Self>) {}
 
     async fn update(mut self, msg: Msg, ctx: EventContext) -> Self {
         match msg {
